Delete generated build code (`configure`, `config.h.in`)
Require users that clone the Git repository to run `autoreconf` to
generate `configure`.  `make dist` includes the generated files in the
release tarball, so users that build from tarball will not need to run
`autoreconf`.

This avoids accidental unsynchronized code (generated code that was
not re-generated after editing an input source) and it will make it
easier to move to Automake.