Support running `autoconf` from a release tarball
Users that download and extract a release tarball will not have a Git
working directory to get the latest tag name from.  Use
`git-version-gen` (copied from gnulib) to support this use case.

Mostly this change exists to enable `make distcheck` when a future
commit migrates the build system to Automake.